Move default page logic to /data and /data/ds
